Split RscCompat into multiple files (fix #146)
