+++ scalafix/rules/src/main/scala/scalafix/internal/rule/semantics/package.scala
@@ -1,0 +1,5 @@
+// Copyright (c) 2017-2018 Twitter, Inc.
+// Licensed under the Apache License, Version 2.0 (see LICENSE.md).
+package scalafix.internal.rule
+
+package object semantics extends Symbols
